Accept multiple file arguments for trash and restore
Accepting multiple file arguments for trash and restore will make
using the command much more convenient for multiple file operations.

Allowing multiple files for restore required adding a new option for
the optional destination directory, rather than making the destination
an optional argument, because with multiple files, it becomes
ambiguous as to whether the last argument is a destination or a file
to restore.